compatibilidad hacia atrás con logger viejo
