+++ config.ts
@@ -43,6 +43,12 @@ export const modules = {
         route: '/core/log',
         middleware: appMiddleware
     },
+    logv2: {
+        active: true,
+        path: './core/log.v2/routes',
+        route: '/core/log.v2',
+        middleware: appMiddleware
+    },
     status: {
         active: true,
         path: './core/status/routes',
