+++ core/log/routes/log.ts
@@ -1,61 +1,48 @@
-import { geoKey } from './../../../config.private';
 import * as express from 'express';
-import { model } from '../schemas/log';
-import { log } from '../controller/log';
-import * as utils from '../../../utils/utils';
+import { Logger } from '../../../utils/logService';
+import { log } from '../schemas/log';
 import { Auth } from '../../../auth/auth.class';
-import { defaultLimit, maxLimit } from './../../../config';
 
 let router = express.Router();
 
-router.post('/', function (req, res, next) {
-    // if (!Auth.check(req, 'log:post')) {
-    //     return next(403);
-    // }
-    if (!req.body.key && !req.body.paciente) {
-        return next(400);
+router.post('/operaciones/:module/:op', function (req, res, next) {
+    if (!Auth.check(req, 'log:post')) {
+        return next(403);
     }
-
-    log(req, req.body.key, req.body.paciente, req.body.operacion, req.body.valor, req.body.anterior).then(
-        (data) => res.json(data),
-        (err) => next(err)
-    );
+    let resultado = Logger.log(req, req.params.module, req.params.op, req.body.data, function (err) {
+        if (err) {
+            return next(err);
+        }
+        res.json(resultado);
+    });
 });
 
-router.get('/', function (req, res, next) {
-    // if (!Auth.check(req, 'log:get')) {
-    //     return next(403);
-    // }
-    if (!req.query.key && !req.query.paciente) {
-        return next(400);
+router.get('/operaciones/:module?', function (req, res, next) {
+    if (!Auth.check(req, 'log:get')) {
+        return next(403);
     }
+    let query;
 
-    let query = model.find({});
-    // Opciones de búsqueda
-    if (req.query.key) {
-        if (req.query.keyRegEx) {
-            query.where('key').regex(new RegExp(req.query.key, 'g'));
-        } else {
-            query.where('key').equals(req.query.key);
-        }
-        query.sort({ key: 1, fecha: -1 });
-    }
-    if (req.query.paciente) {
-        query.where('paciente').equals(req.query.paciente);
-        query.sort({ paciente: 1, fecha: -1 });
-    }
-    if (req.query.desde) {
-        query.where('fecha').gte(req.query.desde);
-    }
-    if (req.query.hasta) {
-        query.where('fecha').gte(req.query.hasta);
-    }
+    query = log.find({});
 
-    // Paginado
-    query.skip(parseInt(req.query.skip || 0, 10));
-    query.limit(Math.min(parseInt(req.query.limit || defaultLimit, 15), maxLimit));
-
-    // Ejecuta la consulta
+    if (req.params.module) {
+        query.where('modulo').equals(req.params.module).sort({ fecha: -1 });
+    }
+    if (req.query.op) {
+        query.where('operacion').equals(req.query.op);
+    }
+    if (req.query.usuario) {
+        query.where('usuario.username').equals(req.query.usuario);
+    }
+    if (req.query.organizacion) {
+        query.where('organizacion._id').equals(req.query.organizacion);
+    }
+    if (req.query.accion) {
+        query.where('datosOperacion.accion').equals(req.query.accion);
+    }
+    if (req.query.count) {
+        query.count();
+    }
     query.exec((err, data) => {
         if (err) {
             return next(err);
