+++ core/log/routes/logPaciente.ts
