+++ core/log/schemas/log.ts
@@ -1,55 +1,67 @@
-import { paciente } from './../../mpi/schemas/paciente';
+// import { mpi } from './../../../config';
 import * as mongoose from 'mongoose';
 import * as organizacion from '../../../core/tm/schemas/organizacion';
 
-export let schema = new mongoose.Schema({
-    key: {
-        type: String,
-        required: function () {
-            return !this.paciente;
-        }
-    },
-    paciente: {
-        type: mongoose.Schema.Types.ObjectId,
-        required: function () {
-            return !this.key;
-        }
-    },
+/**
+ * Descripcion Operaciones logs:
+ * MPI:
+ * scan == escaneo exitoso
+ * scanFail == escaneo fallido
+ * macheoAlto == Macheo con un % superior a 90
+ * posibleDuplicado == Nuevo paciente, pero matchea con un porcentaje entre 80 y 90 con otro.
+ * validadoScan == 'Paciente encontrado por el string del scan'
+ *
+ */
+
+export let logSchema = new mongoose.Schema({
     fecha: {
-        type: Date,
-        required: true,
+        type: Date
+    },
+    usuario: {
+        nombreCompleto: String,
+        nombre: String,
+        apellido: String,
+        username: Number,
+        documento: Number
+    },
+    organizacion: { type: organizacion },
+    modulo: {
+        type: String,
+        enum: ['mpi', 'citas', 'rup', 'configTipoPrestacion', 'fa_sintys', 'fa_sisa', 'fa_anses', 'fa_renaper', 'usuarios', 'scheduler']
     },
     operacion: {
         type: String,
-        required: true
+        enum: [
+            // Operaciones genéricas
+            'query', 'insert', 'update', 'delete',
+            // Operaciones de módulos
+            // ... Mpi
+            'macheoAlto', 'posibleDuplicado', 'reportarError', 'validadoScan', 'scan', 'scanFail',
+            // OperacionesElastic
+            'elasticInsert', 'elasticInsertInPut', 'elasticUpdate', 'elasticDelete', 'elasticError',
+            // ... Citas
+            'asignarTurno', 'cancelarTurno', 'listaEspera', 'actualizarTiposDeTurno', 'actualizarEstadoAgendas', 'actualizarTurnosDelDia', 'lista espera',
+            // ... RUP
+            // hudsPantalla -> Profesional entra a la pantalla de huds de un paciente
+            // hudsPrestacion -> Profesional abre una pretación para ver en la pantalla de ver huds
+            'pacientes', 'hudsPantalla', 'hudsPrestacion',
+            // ...Fuentes Autenticas
+            'validar', 'error',
+            // Scheduler
+            'cda'
+        ]
     },
-    usuario: mongoose.Schema.Types.Mixed,       //  = Se copia idéntico desde el token
-    app: mongoose.Schema.Types.Mixed,           //  = Se copia idéntico desde el token
-    organizacion: mongoose.Schema.Types.Mixed,  //  = Se copia idéntico desde el token
-    data: {
-        anterior: mongoose.Schema.Types.Mixed,
-        valor: mongoose.Schema.Types.Mixed,
-    },
+    datosOperacion: mongoose.Schema.Types.Mixed,
     cliente: {
         ip: String,
-        userAgent: { // utiliza plugin https://github.com/biggora/express-useragent
-            isMobile: Boolean,
-            isDesktop: Boolean,
-            isBot: Boolean,
-            browser: String,
-            version: String,
-            os: String,
-            platform: String,
-            source: String
+        app: {
+            type: String,
+            enum: ['desktop', 'mobile']
         }
     },
     servidor: {
-        ip: String
+        ip: String,
     }
 });
 
-// Indices
-schema.index({ key: 1, fecha: -1 });
-schema.index({ paciente: 1, fecha: -1 });
-
-export let model = mongoose.model('log', schema, 'log');
+export let log = mongoose.model('log', logSchema, 'log');
