+++ core/log/schemas/logAgendaHPNCache.ts
