+++ core/log/schemas/logPaciente.ts
